Add files via upload
Additionally, we implemented a Rolling Window methodology, \autoref{fig:rolling_window}, to train the model with different window sizes (100, 150, 190, 250, 360, and 720 samples). This approach allows the generation of prediction intervals that vary over time and enhances the model's ability to adapt to changes in the transport process.

In the described Rolling Window process, the entire dataset shifts to the right: the training data, the calibration data, and the test data. In each iteration, the rolling window shifts all these sets by the amount defined by window size. This means that at each step, the training and calibration sets move to the right, covering a new portion of historical data, and leaving behind previously covered data, while maintaining their relative position. Finally, the test set shifts as well to cover the subsequent portion of data that will be tested.